Populating manage members table from db
