+++ client/Routes.js
@@ -4,7 +4,7 @@ import Login from './src/screens/Login'
 import CodeValidations from './src/screens/CodeValidations'
 import Settings from './src/screens/Settings'
 import AccountBranding from './src/screens/AccountBranding'
-import ManageTeams from './src/screens/ManageTeams.v1'
+import ManageTeams from './src/screens/ManageTeams'
 import Footer from './src/components/Footer'
 import NavBar from './src/components/NavBar'
 import * as ROUTES from './src/constants/routes'
