+++ client/src/store/index.js
@@ -3,13 +3,13 @@ import { firebase } from '../components/Firebase'
 
 const User = types
   .model({
-    uid: types.string,
+    uuid: types.string,
     isAdmin: types.boolean,
     isSuperAdmin: types.boolean,
+    isActive: types.boolean,
     prefix: types.string,
     firstName: types.string,
     lastName: types.string,
-    role: types.string,
     organizationID: types.string
   });
 
@@ -17,7 +17,8 @@ const Organization = types
   .model({
     name: types.string,
     diagnosisText: types.string,
-    exposureText: types.string
+    exposureText: types.string,
+    members: types.array(User)
   })
 
 const Store = types
@@ -35,12 +36,15 @@ const Store = types
         const userDoc = yield firebase.getUserDocument(email)
 
         self.user = userDoc
-        console.log("user: ", self.user)
 
         const orgDoc = yield firebase.getOrganizationDocument(self.user.organizationID)
 
         self.organization = orgDoc
-        console.log("organization: ", self.organization)
+
+        if (self.user.isAdmin) {
+          const members = yield firebase.getMembersOfOrg(self.user.organizationID)
+          self.organization.members = members
+        }
       } catch (e) {
         console.log(e)
       }
